moved player to utils
